fix: targetmanager reconciles again
Removing the reset of the times was wrong, as the timers only run once.
Replacing with tickers is not possible with the current implementation, as
a ticker must have a positive duration, which would require us to refactor how
the reconcile functions works.

This commit adds more tests that check the proper behavior of the targetmanager
and how many times each call is made. This solved a bug where register would be called
multiple times

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>