Remove unused errors (#121)
* refactor: remove error and context from check shutdown

The error and context values were never used or being produced.
This code-smell was reported in https://github.com/caas-team/sparrow/issues/37
and is solved via this commit.
No need for dead code.

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>

* chore: removed error from unregister check

No error was ever produced, so it can be removed. No need for dead code in the codebase

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>

* chore: removed unneeded errcheck

* chore: registration should be info

The registration of an instance happens once every blue moon, so it should be at least logged at an info level.

---------

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>