test: cleanup on controller_test.go
Added a check for the reconcile method to actually check if the number of check's registered actually matches the expectation and fixed an error in the configuration of a check, which wasn't being registered.

Removed the duplicate check registration case, because this is currently impossible to occur, as the Reconcile function already checks if a check is registered.