feat: Improve traceroute
Adds a couple of features to the traceroute check:
- address of every hop
- dns name, if available, of the hops
- latency for that hop

There have also been major performance improvements:
- All targets are now checked concurrently
- every hop is checked concurrently

So now, instead of waiting for the slowest hop of every target in
series, we parallelized as much as possible. We should now only have to
wait for the slowest target + dns resolution. In my testing (with a
local setup), the time per run with a few target was cut down from a minute-ish
to a few seconds.

Signed-off-by: Niklas Treml <treml.niklas@gmail.com>