fix: removed targets do not produce metrics anymore
A check that gets updated with new targets will not continue exposing the last metric
for that target. The metric(s) with the target label will be deleted.

The responsibility for the metric cleanup must lie with the checks themselves,
as they are the only ones who have access to the metrics themselves.

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>