Fix/remove labelled metrics (#180)
* fix: removed targets do not produce metrics anymore

A check that gets updated with new targets will not continue exposing the last metric
for that target. The metric(s) with the target label will be deleted.

The responsibility for the metric cleanup must lie with the checks themselves,
as they are the only ones who have access to the metrics themselves.

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>

* chore: removed deprecated latency metric

The metric was being already deleted. Instead of just removing the deletion, the metric itself can be finally removed for the next version.

* chore: clean up & renaming

Renamed the parameter of the metric cleanup to make the behavior clearer

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>

* chore: clearer comment

The term Collector for the prometheus metrics is avoided, as it was only creating confusion

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>

* fix: removed double increase of count

The metrics was being incremented twice for each check. Now it is not.

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>

* chore: use quoted string for error

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>

* chore: restructuring for consistency

---------

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>
Co-authored-by: lvlcn-t <75443136+lvlcn-t@users.noreply.github.com>